Added Example Scene for Outline.
I added an example scene for the Outline effect. The artists need to create an inverted model where the normals of the mesh are inverted. This will allow us to add an unlit shader onto the model and scale it ever so slightly so that it creates an outline effect.